+++ source/core/Main.js
@@ -125,7 +125,6 @@ export {ParticleEmitter} from "./objects/particle/ParticleEmitter.js";
 export {Sky} from "./objects/misc/Sky.js";
 export {Group} from "./objects/misc/Group.js";
 export {LensFlare} from "./objects/misc/LensFlare.js";
-export {HTMLView} from "./objects/misc/HTMLView.js";
 
 export {OrbitControls} from "./objects/controls/OrbitControls.js";
 export {FirstPersonControls} from "./objects/controls/FirstPersonControls.js";